// apps/desktop/src/convex/_generated/api.d.ts
/* eslint-disable */
/**
 * Generated `api` utility.
 *
 * THIS CODE IS AUTOMATICALLY GENERATED.
 *
 * To regenerate, run `npx convex dev`.
 * @module
 */

import type * as accounts from "../accounts.js";
import type * as budgets from "../budgets.js";
import type * as paymentMethods from "../paymentMethods.js";
import type * as payments from "../payments.js";
import type * as pushSubscriptions from "../pushSubscriptions.js";
import type * as reminders from "../reminders.js";
import type * as subscriptions from "../subscriptions.js";
import type * as templates from "../templates.js";
import type * as transactions from "../transactions.js";
import type * as userSettings from "../userSettings.js";

import type {
  ApiFromModules,
  FilterApi,
  FunctionReference,
} from "convex/server";

declare const fullApi: ApiFromModules<{
  accounts: typeof accounts;
  budgets: typeof budgets;
  paymentMethods: typeof paymentMethods;
  payments: typeof payments;
  pushSubscriptions: typeof pushSubscriptions;
  reminders: typeof reminders;
  subscriptions: typeof subscriptions;
  templates: typeof templates;
  transactions: typeof transactions;
  userSettings: typeof userSettings;
}>;

/**
 * A utility for referencing Convex functions in your app's public API.
 *
 * Usage:
 * ```js
 * const myFunctionReference = api.myModule.myFunction;
 * ```
 */
export declare const api: FilterApi<
  typeof fullApi,
  FunctionReference<any, "public">
>;

/**
 * A utility for referencing Convex functions in your app's internal API.
 *
 * Usage:
 * ```js
 * const myFunctionReference = internal.myModule.myFunction;
 * ```
 */
export declare const internal: FilterApi<
  typeof fullApi,
  FunctionReference<any, "internal">
>;